+++ internal/command/darwin.go
@@ -51,7 +51,7 @@ func (cmd *Darwin) Parse(args []string) error {
 	flagSet.Var(flags.TargetArch, "arch", fmt.Sprintf(`List of target architecture to build separated by comma. Supported arch: %s`, darwinArchSupported))
 
 	// flags used only in release mode
-	flagSet.StringVar(&flags.Category, "category", "", "The name of the certificate to sign the build")
+	flagSet.StringVar(&flags.Category, "category", "", "The category of the app for store listing")
 
 	flagAppID := flagSet.Lookup("app-id")
 	flagAppID.Usage = fmt.Sprintf("%s [required]", flagAppID.Usage)
@@ -113,7 +113,7 @@ func (cmd *Darwin) Run() error {
 				return fmt.Errorf("darwin release build is supported only on darwin hosts")
 			}
 
-			packageName = fmt.Sprintf("%s.pkg", ctx.Output)
+			packageName = fmt.Sprintf("%s.pkg", ctx.Name)
 			srcFile = volume.JoinPathHost(ctx.WorkDirHost(), packageName)
 
 			err = fyneReleaseHost(ctx)
@@ -126,7 +126,7 @@ func (cmd *Darwin) Run() error {
 				return err
 			}
 
-			packageName = fmt.Sprintf("%s.app", ctx.Output)
+			packageName = fmt.Sprintf("%s.app", ctx.Name)
 			srcFile = volume.JoinPathHost(ctx.TmpDirHost(), ctx.ID, packageName)
 
 			err = fynePackage(ctx)