+++ internal/command/linux.go
@@ -106,7 +106,7 @@ func (cmd *Linux) Run() error {
 		//
 		log.Info("[i] Packaging app...")
 
-		packageName := fmt.Sprintf("%s.tar.gz", ctx.Output)
+		packageName := fmt.Sprintf("%s.tar.gz", ctx.Name)
 
 		err = prepareIcon(ctx)
 		if err != nil {